+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Fixed:
+- Change contract update cost calculation to local computation instead of relying on the wallet proxy.
 - Fixed wallet import being blocked by an incorrect Terms & Conditions acceptance prompt after the user had already accepted it
 
 ## [Released 3.5.1]